Allow arbitrary PDB path padding in CodeView
