+++ src/System.Reflection.Metadata/tests/PortableExecutable/DebugDirectoryBuilderTests.cs
@@ -278,114 +278,5 @@ namespace System.Reflection.PortableExecutable.Tests
                 }
             }
         }
-
-        [Fact]
-        public void EmbeddedPortablePdb_Errors()
-        {
-            var bytes1 = ImmutableArray.Create(new byte[]
-            {
-                0x4D, 0x50, 0x44, 0x42, // signature
-                0xFF, 0xFF, 0xFF, 0xFF, // uncompressed size
-                0xEB, 0x28, 0x4F, 0x0B, 0x75, 0x31, 0x56, 0x12, 0x04, 0x00 // compressed data
-            });
-
-            using (var block = new ByteArrayMemoryProvider(bytes1).GetMemoryBlock(0, bytes1.Length))
-            {
-                Assert.Throws<BadImageFormatException>(() => PEReader.DecodeEmbeddedPortablePdbDebugDirectoryData(block));
-            }
-
-            var bytes2 = ImmutableArray.Create(new byte[]
-            {
-                0x4D, 0x50, 0x44, 0x42, // signature
-                0x09, 0x00, 0x00, 0x00, // uncompressed size
-                0xEB, 0x28, 0x4F, 0x0B, 0x75, 0x31, 0x56, 0x12, 0x04, 0x00 // compressed data
-            });
-
-            using (var block = new ByteArrayMemoryProvider(bytes2).GetMemoryBlock(0, bytes2.Length))
-            {
-                Assert.Throws<BadImageFormatException>(() => PEReader.DecodeEmbeddedPortablePdbDebugDirectoryData(block));
-            }
-
-            var bytes3 = ImmutableArray.Create(new byte[]
-            {
-                0x4D, 0x50, 0x44, 0x42, // signature
-                0x00, 0x00, 0x00, 0x00, // uncompressed size
-                0xEB, 0x28, 0x4F, 0x0B, 0x75, 0x31, 0x56, 0x12, 0x04, 0x00 // compressed data
-            });
-
-            using (var block = new ByteArrayMemoryProvider(bytes3).GetMemoryBlock(0, bytes3.Length))
-            {
-                Assert.Throws<BadImageFormatException>(() => PEReader.DecodeEmbeddedPortablePdbDebugDirectoryData(block));
-            }
-
-            var bytes4 = ImmutableArray.Create(new byte[]
-            {
-                0x4D, 0x50, 0x44, 0x42, // signature
-                0xff, 0xff, 0xff, 0x7f, // uncompressed size
-                0xEB, 0x28, 0x4F, 0x0B, 0x75, 0x31, 0x56, 0x12, 0x04, 0x00 // compressed data
-            });
-
-            using (var block = new ByteArrayMemoryProvider(bytes4).GetMemoryBlock(0, bytes4.Length))
-            {
-                Assert.Throws<BadImageFormatException>(() => PEReader.DecodeEmbeddedPortablePdbDebugDirectoryData(block));
-            }
-
-            var bytes5 = ImmutableArray.Create(new byte[]
-            {
-                0x4D, 0x50, 0x44, 0x42, // signature
-                0x08, 0x00, 0x00, 0x00, // uncompressed size
-                0xEF, 0xFF, 0x4F, 0xFF, 0x75, 0x31, 0x56, 0x12, 0x04, 0x00 // compressed data
-            });
-
-            using (var block = new ByteArrayMemoryProvider(bytes4).GetMemoryBlock(0, bytes4.Length))
-            {
-                Assert.Throws<BadImageFormatException>(() => PEReader.DecodeEmbeddedPortablePdbDebugDirectoryData(block));
-            }
-
-            var bytes6 = ImmutableArray.Create(new byte[]
-            {
-                0x4D, 0x50, 0x44, 0x43, // signature
-                0x08, 0x00, 0x00, 0x00, // uncompressed size
-                0xEB, 0x28, 0x4F, 0x0B, 0x75, 0x31, 0x56, 0x12, 0x04, 0x00 // compressed data
-            });
-
-            using (var block = new ByteArrayMemoryProvider(bytes6).GetMemoryBlock(0, bytes6.Length))
-            {
-                Assert.Throws<BadImageFormatException>(() => PEReader.DecodeEmbeddedPortablePdbDebugDirectoryData(block));
-            }
-
-            var bytes7 = ImmutableArray.Create(new byte[]
-            {
-                0x4D, 0x50, 0x44, 0x43, // signature
-                0x08, 0x00, 0x00, 
-            });
-
-            using (var block = new ByteArrayMemoryProvider(bytes7).GetMemoryBlock(0, bytes7.Length))
-            {
-                Assert.Throws<BadImageFormatException>(() => PEReader.DecodeEmbeddedPortablePdbDebugDirectoryData(block));
-            }
-
-            var bytes8 = ImmutableArray.Create(new byte[]
-            {
-                0x4D, 0x50, 0x44, 0x43, // signature
-                0x08, 0x00, 0x00,
-            });
-
-            using (var block = new ByteArrayMemoryProvider(bytes8).GetMemoryBlock(0, bytes8.Length))
-            {
-                Assert.Throws<BadImageFormatException>(() => PEReader.DecodeEmbeddedPortablePdbDebugDirectoryData(block));
-            }
-
-            var bytes9 = ImmutableArray.Create(new byte[]
-            {
-                0x4D, 0x50, 0x44, 0x43, // signature
-                0x08, 0x00, 0x00, 0x00
-            });
-
-            using (var block = new ByteArrayMemoryProvider(bytes9).GetMemoryBlock(0, 1))
-            {
-                Assert.Throws<BadImageFormatException>(() => PEReader.DecodeEmbeddedPortablePdbDebugDirectoryData(block));
-            }
-        }
     }
 }
